# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API for NutriApp - Nutritionist management platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /groups:
    get:
      operationId: listGroups
      tags: [groups]
      summary: List groups
      responses:
        "200":
          description: List of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"
    post:
      operationId: createGroup
      tags: [groups]
      summary: Create group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupInput"
      responses:
        "201":
          description: Created group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
  /groups/{id}:
    get:
      operationId: getGroup
      tags: [groups]
      summary: Get group with members
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Group detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupDetail"
        "404":
          description: Not found
    patch:
      operationId: updateGroup
      tags: [groups]
      summary: Update group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupUpdate"
      responses:
        "200":
          description: Updated group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
    delete:
      operationId: deleteGroup
      tags: [groups]
      summary: Delete group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /groups/{id}/members:
    post:
      operationId: addGroupMember
      tags: [groups]
      summary: Add patient to group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupMemberInput"
      responses:
        "201":
          description: Member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
  /groups/{id}/members/{patientId}:
    delete:
      operationId: removeGroupMember
      tags: [groups]
      summary: Remove patient from group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # ── Consultations ──────────────────────────────────────
  /patients/{patientId}/consultations:
    get:
      operationId: listConsultations
      tags: [consultations]
      summary: List patient consultations
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Consultations list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Consultation"
    post:
      operationId: createConsultation
      tags: [consultations]
      summary: Create consultation
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsultationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultation"
  /patients/{patientId}/consultations/{consultationId}:
    get:
      operationId: getConsultation
      tags: [consultations]
      summary: Get consultation detail
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
        - name: consultationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Consultation detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsultationDetail"
        "404":
          description: Not found
    patch:
      operationId: updateConsultation
      tags: [consultations]
      summary: Update consultation
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
        - name: consultationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsultationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsultationDetail"
    delete:
      operationId: deleteConsultation
      tags: [consultations]
      summary: Delete consultation
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
        - name: consultationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Biblioteca (admin) ─────────────────────────────────
...
components:
  schemas:
...
    SuccessMessage:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ── Auth Schemas ──────────────────────────────────────
...
    Patient:
      type: object
      required: [id, name, email, createdAt, hasAccess]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        birthDate:
          type: ["string", "null"]
        sex:
          type: ["string", "null"]
        hasAccess:
          type: boolean
        createdAt:
          type: string
        groups:
          type: array
          items:
            $ref: "#/components/schemas/GroupSummary"
...
    GroupSummary:
      type: object
      required: [id, name, color]
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
    Group:
      type: object
      required: [id, name, color, createdAt, memberCount]
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
        memberCount:
          type: integer
    GroupDetail:
      type: object
      required: [id, name, color, createdAt, members]
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
        members:
          type: array
          items:
            $ref: "#/components/schemas/Patient"
    GroupInput:
      type: object
      required: [name, color]
      properties:
        name:
          type: string
          minLength: 1
        color:
          type: string
        description:
          type: string
    GroupUpdate:
      type: object
      properties:
        name:
          type: string
        color:
          type: string
        description:
          type: ["string", "null"]
    GroupMemberInput:
      type: object
      required: [patientId]
      properties:
        patientId:
          type: integer

    # ── Consultation Schemas ───────────────────────────────
    Consultation:
      type: object
      required: [id, patientId, type, date, createdAt]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        type:
          type: string
          enum: [avaliacao_inicial, cardapio, reavaliacao]
        date:
          type: string
        weight:
          type: ["number", "null"]
        height:
          type: ["number", "null"]
        bmi:
          type: ["number", "null"]
        createdAt:
          type: string
    ConsultationDetail:
      type: object
      required: [id, patientId, type, date, createdAt]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        type:
          type: string
          enum: [avaliacao_inicial, cardapio, reavaliacao]
        date:
          type: string
        createdAt:
          type: string
        # Avaliação Inicial
        weight:
          type: ["number", "null"]
        height:
          type: ["number", "null"]
        bmi:
          type: ["number", "null"]
        waistCm:
          type: ["number", "null"]
        hipCm:
          type: ["number", "null"]
        abdomenCm:
          type: ["number", "null"]
        armCm:
          type: ["number", "null"]
        thighCm:
          type: ["number", "null"]
        calfCm:
          type: ["number", "null"]
        bioimpedancePdfPath:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        # Cardápio
        objective:
          type: ["string", "null"]
        vetKcal:
          type: ["number", "null"]
        menuPdfPath:
          type: ["string", "null"]
        restrictions:
          type: ["string", "null"]
        additionalGuidance:
          type: ["string", "null"]
        # Reavaliação
        baselineConsultationId:
          type: ["integer", "null"]
        evolutionNotes:
          type: ["string", "null"]
        comparison:
          $ref: "#/components/schemas/ConsultationComparison"
    ConsultationComparison:
      type: object
      properties:
        weightDelta:
          type: ["number", "null"]
        bmiDelta:
          type: ["number", "null"]
        waistDelta:
          type: ["number", "null"]
        hipDelta:
          type: ["number", "null"]
        abdomenDelta:
          type: ["number", "null"]
        armDelta:
          type: ["number", "null"]
        thighDelta:
          type: ["number", "null"]
        calfDelta:
          type: ["number", "null"]
    ConsultationInput:
      type: object
      required: [type, date]
      properties:
        type:
          type: string
          enum: [avaliacao_inicial, cardapio, reavaliacao]
        date:
          type: string
        # Avaliação Inicial / Reavaliação
        weight:
          type: ["number", "null"]
        height:
          type: ["number", "null"]
        waistCm:
          type: ["number", "null"]
        hipCm:
          type: ["number", "null"]
        abdomenCm:
          type: ["number", "null"]
        armCm:
          type: ["number", "null"]
        thighCm:
          type: ["number", "null"]
        calfCm:
          type: ["number", "null"]
        bioimpedancePdfPath:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        # Cardápio
        objective:
          type: ["string", "null"]
        vetKcal:
          type: ["number", "null"]
        menuPdfPath:
          type: ["string", "null"]
        restrictions:
          type: ["string", "null"]
        additionalGuidance:
          type: ["string", "null"]
        # Reavaliação
        baselineConsultationId:
          type: ["integer", "null"]
        evolutionNotes:
          type: ["string", "null"]
    ConsultationUpdate:
      type: object
      properties:
        date:
          type: string
        weight:
          type: ["number", "null"]
        height:
          type: ["number", "null"]
        waistCm:
          type: ["number", "null"]
        hipCm:
          type: ["number", "null"]
        abdomenCm:
          type: ["number", "null"]
        armCm:
          type: ["number", "null"]
        thighCm:
          type: ["number", "null"]
        calfCm:
          type: ["number", "null"]
        bioimpedancePdfPath:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        objective:
          type: ["string", "null"]
        vetKcal:
          type: ["number", "null"]
        menuPdfPath:
          type: ["string", "null"]
        restrictions:
          type: ["string", "null"]
        additionalGuidance:
          type: ["string", "null"]
        baselineConsultationId:
          type: ["integer", "null"]
        evolutionNotes:
          type: ["string", "null"]

    # ── Publication Schemas ───────────────────────────────
...