# lib/api-zod/src/generated/types/groupInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API for NutriApp - Nutritionist management platform
 * OpenAPI spec version: 0.1.0
 */

export interface GroupInput {
  /** @minLength 1 */
  name: string;
  color: string;
  description?: string;
}